Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
rembg
pillow
onnxruntime